Adds API to import encoded private keys (#4357)
* Adds API to import encoded private keys

* Adds solana base58 decoding as well

* Supports 64 byte solana key import and moves Crc

* Adds API to retrieve encoded key

* Adds APIs and Tests based on the discussion

* Addresses review comments

* Addresses review comments

* Updates Android test

+++ wasm/src/keystore/default-impl.ts
@@ -104,6 +104,26 @@ export class Default implements Types.IKeyStore {
     });
   }
 
+  importKeyEncoded(
+    key: string,
+    name: string,
+    password: string,
+    coin: CoinType,
+    encryption: StoredKeyEncryption
+  ): Promise<Types.Wallet> {
+    return new Promise((resolve, reject) => {
+      const { StoredKey, PrivateKey, Curve, StoredKeyEncryption } = this.core;
+
+      let pass = Buffer.from(password);
+      let storedKey = StoredKey.importPrivateKeyEncodedWithEncryption(key, name, pass, coin, encryption);
+      let wallet = this.mapWallet(storedKey);
+      storedKey.delete();
+      this.importWallet(wallet)
+        .then(() => resolve(wallet))
+        .catch((error) => reject(error));
+    });
+  }
+
   addAccounts(
     id: string,
     password: string,
@@ -160,6 +180,46 @@ export class Default implements Types.IKeyStore {
         default:
           throw Types.Error.InvalidJSON;
       }
+      storedKey.delete();
+      return value;
+    });
+  }
+
+  getWalletType(id: string): Promise<Types.WalletType> {
+    return this.load(id).then((wallet) => wallet.type);
+  }
+
+  exportMnemonic(id: string, password: string): Promise<string> {
+    return this.load(id).then((wallet) => {
+      if (wallet.type !== Types.WalletType.Mnemonic) {
+        throw Types.Error.UnsupportedWalletType;
+      }
+      let storedKey = this.mapStoredKey(wallet);
+      let value = storedKey.decryptMnemonic(Buffer.from(password));
+      storedKey.delete();
+      return value;
+    });
+  }
+
+  exportPrivateKey(id: string, password: string): Promise<Uint8Array> {
+    return this.load(id).then((wallet) => {
+      if (wallet.type !== Types.WalletType.PrivateKey) {
+        throw Types.Error.UnsupportedWalletType;
+      }
+      let storedKey = this.mapStoredKey(wallet);
+      let value = storedKey.decryptPrivateKey(Buffer.from(password));
+      storedKey.delete();
+      return value;
+    });
+  }
+
+  exportPrivateKeyEncoded(id: string, password: string): Promise<string> {
+    return this.load(id).then((wallet) => {
+      if (wallet.type !== Types.WalletType.PrivateKey) {
+        throw Types.Error.UnsupportedWalletType;
+      }
+      let storedKey = this.mapStoredKey(wallet);
+      let value = storedKey.decryptPrivateKeyEncoded(Buffer.from(password));
       storedKey.delete();
       return value;
     });
